+++ src/main/java/com/andersen_intensive/hotel/service/UtilityService.java
@@ -2,10 +2,11 @@ package com.andersen_intensive.hotel.service;
 
 import com.andersen_intensive.hotel.models.Utility;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public interface UtilityService {
-    Utility saveService(String name, double price);
+    Utility saveService(String name, BigDecimal price);
     List<Utility> sortByName();
     List<Utility> sortByPrice();
 }
